chat-app-backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
alembic
pydantic
pydantic-settings
python-dotenv
aioredis
itsdangerous
pywebpush
requests
beautifulsoup4
slowapi
python-multipart
boto3
fastapi-cache2[redis]